requirements.txt
streamlit
requests
jsonschema
basyx-python-sdk==1.2.1